AirflowETLService/requirements.txt
apache-airflow>=2.7.0
apache-airflow-providers-postgres
psycopg2-binary
pandas>=2.1.3
numpy>=1.26.2
